requirements.txt
flask
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy
sqlalchemy
email_validator
python-decouple
gunicorn
jinja2
pyopenssl
werkzeug
pyjks
